// client/src/app/models/custom-domain.model.ts
/**
 * CustomDomainModel is holding the data param for each request like 
 * `create`, `list`, `update` and `delete` custom domain.
 * @class CustomDomainModel
 * @version 1.0.0
 * @author Tier5 LLC `<work@tier5.us>`
 * @license Proprietary
 */
export class CustomDomainModel {
  name?: string;
  domain?: string;
  custom_domain_id?: string;
  id?: string;
}